pyproject.toml
[tool.poetry]
name = "payhere"
version = "0.1.0"
description = ""
authors = ["None"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "3.2.13"
djangorestframework = "3.13.1"
ipython = "8.4.0"
python-dotenv = "0.20.0"
django-extensions = "3.1.5"
drf-yasg = "1.20.0"
PyMySQL = "1.0.2"
mysqlclient = "2.1.1"
djangorestframework-simplejwt = "5.2.0"
bcrypt = "3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
